client/requirements.txt
pip==21.3.1
streamlit==1.20.0
requests
python-dotenv
extra-streamlit-components
streamlit-extras
pandas
altair
gunicorn==20.1.0

